worked in comments and ratings
