+++ src/features/usershoppinghistory.js
@@ -3,7 +3,7 @@ import { createAction, createReducer } from "@reduxjs/toolkit";
 const addtousershoppinghistory = createAction('is adding to shopping history');
 const resetusershoppinghistory = createAction('is resetting to shopping history');
 
-const actionsshoppinghistory = {addtousershoppinghistory};
+const actionsshoppinghistory = {addtousershoppinghistory, resetusershoppinghistory};
 
 const initialState = {
     currentusershoppinghistory: []
